--- rest-api/REST-API/REST-API/src/main/java/deployment/Application.java
@@ -1,25 +1,0 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-package deployment;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
-import org.springframework.context.annotation.ComponentScan;
-import org.springframework.context.annotation.Configuration;
-
-/**
- *
- * @author Pedro
- */
-@Configuration
-@EnableAutoConfiguration
-@ComponentScan({"controllers","services"})
-
-public class Application {
-    public static void main(String[] args){
-        SpringApplication.run(Application.class, args);
-    }
-}
