--- rest-api/REST-API/REST-API/src/main/java/model/LoginBody.java
@@ -1,40 +1,0 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-package model;
-
-/**
- *
- * @author Carlos Soares
- */
-public class LoginBody {
-    private String username;
-    private String password;
-
-    public LoginBody(String username, String password) {
-        this.username = username;
-        this.password = password;
-    }
-  
-    
-
-    public String getUsername() {
-        return username;
-    }
-
-    public void setUsername(String username) {
-        this.username = username;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public void setPassword(String password) {
-        this.password = password;
-    }
-    
-    
-}
