+++ rest-api/REST-API/src/main/java/controllers/ProdutoController.java
@@ -14,6 +14,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import static org.springframework.web.bind.annotation.RequestMethod.GET;
+import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import services.ProdutoService;
 
@@ -35,6 +36,33 @@ public class ProdutoController {
                 new ResponseEntity<List<Product>>(list,HttpStatus.OK): 
                 new ResponseEntity<String>("No products were found",HttpStatus.NOT_FOUND);
     }
+    
+    @RequestMapping(value="/byID",method = GET)
+    public ResponseEntity<?> getProductByID(@RequestParam(value="id",required=false)Integer id){
+        if (id == null) return new  ResponseEntity<String>("ID field wasn't sent",HttpStatus.BAD_REQUEST);
+        Product product = productService.getProductByID(id);
+        return product!=null ? 
+                new ResponseEntity<Product>(product,HttpStatus.OK): 
+                new ResponseEntity<String>("No product with this id",HttpStatus.NOT_FOUND);
+    }
    
+    @RequestMapping(value="/byCategory",method = GET)
+    public ResponseEntity<?> getProductByCategory(@RequestParam(value="category",required=false)String category){
+        if (category == null) return new  ResponseEntity<String>("Category field wasn't sent",HttpStatus.BAD_REQUEST);
+        List<Product> products = productService.getProductByCategory(category);
+        return products!=null && !products.isEmpty() ? 
+                new ResponseEntity<List<Product>>(products,HttpStatus.OK): 
+                new ResponseEntity<String>("No product within this category",HttpStatus.NOT_FOUND);
+    }
 
+    
+    
+    
+    @RequestMapping(value="/categories",method = GET)
+    public ResponseEntity<?> getAllCategories(){
+        List<String> categories = productService.getAllCategories();
+        return categories!=null && !categories.isEmpty() ? 
+                new ResponseEntity<List<String>>(categories,HttpStatus.OK): 
+                new ResponseEntity<String>("No categories available",HttpStatus.NOT_FOUND);
+    }
 }
