+++ rest-api/REST-API/src/main/java/data/MockData.java
@@ -10,6 +10,7 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import model.Product;
+import model.User;
 
 /**
  *
@@ -17,16 +18,17 @@ import model.Product;
  */
 
 public class MockData {
-    private static Map<String,String> users;
+    private static Map<String,String> logins;
     private static List<Product> allProducts;
+    private static List<User> users;
 
     public static Map<String, String> getUsers() {
-        users = new HashMap<>();
-        users.put("salome","123");
-        users.put("arinto","123");
-        users.put("ze","123");
-        users.put("carlos","123");
-        return users;
+        logins = new HashMap<>();
+        logins.put("salome","123");
+        logins.put("arinto","123");
+        logins.put("ze","123");
+        logins.put("carlos","123");
+        return logins;
     }
 
     public static List<Product> getAllProducts() {
@@ -35,9 +37,20 @@ public class MockData {
         allProducts.add(new Product(2,"batata","legumes",1.49f,"é castanha"));
         allProducts.add(new Product(3,"alho","legumes",4.99f,"é mau"));
         allProducts.add(new Product(4,"couve branca","legumes",1.29f,"é verde"));
+        allProducts.add(new Product(5,"morangos","fruta",1.29f,"é vermelho"));
         
         return allProducts;
     }
+    
+    public static List<User> getAllUsers() {
+        users = new ArrayList<>();
+        users.add(new User(1, "salome", "123", "salome lopes", "salome@ua.pt", 785412965, "Rua Direita n85", "3810-895", 915268478));
+        users.add(new User(1, "arinto", "123", "carlos arinto", "arinto@ua.pt",123458967 , "Rua Direita n85", "3810-895", 915268478));
+        users.add(new User(1, "ze", "123", "jose ribeiro", "ze@ua.pt", 852369741, "Rua Direita n85", "3810-895", 915268478));
+        users.add(new User(1, "carlos", "123", "carlos soares", "carlos@ua.pt", 225366987, "Rua Direita n85", "3810-895", 915268478));
+        return users;
+    }
+
