+++ rest-api/REST-API/src/main/java/services/ProdutoService.java
@@ -6,7 +6,9 @@
 package services;
 
 import data.MockData;
+import java.util.ArrayList;
 import java.util.List;
+import java.util.stream.Collectors;
 import model.Product;
 import org.springframework.stereotype.Service;
 
@@ -17,8 +19,24 @@ import org.springframework.stereotype.Service;
 @Service
 public class ProdutoService {
     
-    
     public List<Product> getAllProducts(){
         return MockData.getAllProducts();
+    }
+
+    public Product getProductByID(int id) {
+        Product product = MockData.getAllProducts().stream().filter((f) -> id == f.getId()).findAny().orElse(null);
+        return product;
+    }
+
+    public List<Product> getProductByCategory(String category) {
+        return MockData.getAllProducts().stream().filter(f -> f.getCategory().equals(category)).collect(Collectors.toList());
+    }
+
+    public List<String> getAllCategories() {
+        List<String> categories = new ArrayList<>();
+        MockData.getAllProducts().forEach(f ->{
+            if (!categories.contains(f.getCategory()))
+                categories.add(f.getCategory());});
+        return categories;
     }
 }
